// src/types/dx-multiple-select-chip.ts
/* ======================================================================== *
 * Copyright 2025 HCL America Inc.                                          *
 * Licensed under the Apache License, Version 2.0 (the "License");          *
 * you may not use this file except in compliance with the License.         *
 * You may obtain a copy of the License at                                  *
 *                                                                          *
 * http://www.apache.org/licenses/LICENSE-2.0                               *
 *                                                                          *
 * Unless required by applicable law or agreed to in writing, software      *
 * distributed under the License is distributed on an "AS IS" BASIS,        *
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. *
 * See the License for the specific language governing permissions and      *
 * limitations under the License.                                           *
 * ======================================================================== */

export type DxMultiSelectChip = {
  toggleDropDown: boolean;
  label: string;
  disabled: boolean;
  selectedValues: string[];
  options: string[];
  hiddenLabel: boolean;
  showChipCount: boolean;
  showAvatar: boolean;
  emptyOptions: boolean;
  clearIcon: boolean;
  fullWidth: boolean;
  error: boolean;
}

export type OptionData = {
  id: string;
  name: string;
  value: string;
}

export enum DxMultiSelectInputFieldType {
  MULTI_SELECT_SOURCE = 'multiSelectSource',
  DOCUMENT_OBJECT_TYPE = 'documentObjectType',
  QUERY_STRING = 'queryString',
  TRIGGER_BUTTON = 'triggerButton',
  PAGINATION_ROWS = 'paginationRows',
  PAGINATION_PAGE = 'paginationPage',
  SORT = 'sort',
  CLEAR_QUERY = 'clearQuery',
  SELECTION = 'selection',
  SELECT_ALL = 'selectAll',
}

export type MultiSelectChangeDetail = Array<{
  id: string;
  value: string;
  name: string;
}>;